+++ examples/data-sources/nxos_dhcp/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_dhcp" "example" {
+}