+++ examples/data-sources/nxos_dhcp_relay_interface/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_dhcp_relay_interface" "example" {
-  interface_id = "eth1/10"
-}