+++ examples/resources/nxos_dhcp/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = nxos_dhcp.example
+  id = ""
+}