+++ examples/resources/nxos_dhcp/resource.tf
@@ -1,0 +1,9 @@
+resource "nxos_dhcp" "example" {
+  relay_interfaces = [{
+    interface_id = "eth1/10"
+    addresses = [{
+      vrf     = "VRF1"
+      address = "1.1.1.1"
+    }]
+  }]
+}