+++ examples/resources/nxos_dhcp_relay_address/import-by-identity.tf
@@ -1,8 +1,0 @@
-import {
-  to = nxos_dhcp_relay_address.example
-  identity = {
-    "interface_id" : "<interface_id>"
-    "vrf" : "<vrf>"
-    "address" : "<address>"
-  }
-}