+++ examples/resources/nxos_dhcp_relay_address/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_dhcp_relay_address.example
-  id = "<interface_id>,<vrf>,<address>"
-}