+++ examples/resources/nxos_dhcp_relay_address/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_dhcp_relay_address.example "<interface_id>,<vrf>,<address>"