+++ examples/resources/nxos_dhcp_relay_interface/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_dhcp_relay_interface.example "<interface_id>"