+++ examples/resources/nxos_dhcp_relay_interface/resource.tf
@@ -1,3 +1,0 @@
-resource "nxos_dhcp_relay_interface" "example" {
-  interface_id = "eth1/10"
-}